+++ CHANGELOG.md
@@ -4,6 +4,7 @@ ArduinoJson: change log
 HEAD
 ----
 
+* Fix buffer overrun when reading a raw string with a truncated MessagePack extension header as `MsgPackExtension`
 * Don't store string literals by pointer anymore (issue #2189)
   Version 7.3 introduced a new way to detect string literals, but it fails in some edge cases.
   I could not find a way to fix it, so I chose to remove the optimization rather than keep it broken.